requirements.txt
rauth
requests
setuptools
pyyaml
pandas
xlsxwriter
python-dateutil